create ai base client and cleanup
