+++ sdk/ahnlich-client-py/ahnlich_client_py/__init__.py
@@ -1,4 +1,4 @@
 from ahnlich_client_py import builders, libs
-from ahnlich_client_py.clients import AhnlichDBClient
+from ahnlich_client_py.clients import AhnlichAIClient, AhnlichDBClient
 from ahnlich_client_py.config import AhnlichDBPoolSettings
 from ahnlich_client_py.internals import ai_query, ai_response, db_query, db_response
