+++ sdk/ahnlich-client-py/ahnlich_client_py/builders/ai.py
@@ -1,2 +1,13 @@
+import typing
+
+import numpy as np
+
+from ahnlich_client_py import exceptions as ah_exceptions
+from ahnlich_client_py.internals import ai_query
+from ahnlich_client_py.internals import serde_types as st
+from ahnlich_client_py.libs import NonZeroSizeInteger
+
+
 class AhnlichAIRequestBuilder:
-    pass
+    def __init__(self) -> None:
+        self.queries: typing.List[ai_query.AIQuery] = []
