+++ sdk/ahnlich-client-py/ahnlich_client_py/builders/db.py
@@ -5,16 +5,7 @@ import numpy as np
 from ahnlich_client_py import exceptions as ah_exceptions
 from ahnlich_client_py.internals import db_query
 from ahnlich_client_py.internals import serde_types as st
-
-
-class NonZeroSizeInteger:
-    def __init__(self, num: st.uint64) -> None:
-
-        if num <= 0:
-            raise ah_exceptions.AhnlichValidationError(
-                "Ahnlich expects a Non zero value as integers"
-            )
-        self.value = num
+from ahnlich_client_py.libs import NonZeroSizeInteger
 
 
 class AhnlichDBRequestBuilder:
