+++ sdk/ahnlich-client-py/ahnlich_client_py/clients/__init__.py
@@ -1,1 +1,2 @@
+from .ai import AhnlichAIClient
 from .db import AhnlichDBClient
